+++ src/hud.cpp
@@ -622,7 +622,7 @@ void Hud::resizeHotbar() {
 }
 
 struct MeshTimeInfo {
-	s32 time;
+	s64 time;
 	scene::IMesh *mesh;
 };
